cluster watcher hooks
