+++ plugins/alerting/pkg/alerting/condition_handlers.go
@@ -6,14 +6,13 @@ package alerting
 
 import (
 	"context"
-	"encoding/json"
+	"strings"
 	"sync"
 	"time"
 
 	"github.com/nats-io/nats.go"
 	"github.com/prometheus/common/model"
 	"github.com/rancher/opni/pkg/health"
-	"github.com/rancher/opni/plugins/alerting/pkg/alerting/alertstorage"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/rancher/opni/pkg/alerting/metrics"
@@ -302,153 +301,6 @@ func (p *Plugin) handlePrometheusQueryAlertCreation(ctx context.Context, q *aler
 	return err
 }
 
-type InternalConditionEvaluator[T any] struct {
-	// metadata
-	lg                 *zap.SugaredLogger
-	conditionName      string
-	conditionId        string
-	clusterId          string
-	alertmanagerlabels map[string]string
-	// contexts
-	parentCtx        context.Context
-	cancelEvaluation context.CancelFunc
-	evaluateDuration time.Duration
-	evaluationCtx    context.Context
-
-	inMemoryFiring bool
-	stateLock      sync.Mutex
-	firingLock     sync.RWMutex
-
-	// closure to evaluate health from the subscriber
-	alpha       func(h T) (healthy bool, ts *timestamppb.Timestamp)
-	triggerHook func(ctx context.Context, conditionId string, annotations map[string]string)
-	storageNode *alertstorage.StorageNode
-	msgCh       chan *nats.Msg
-}
-
-// infinite & blocking : must be run in a goroutine
-func (c *InternalConditionEvaluator[T]) SubscriberLoop() {
-	defer c.cancelEvaluation()
-	if c.msgCh == nil { // TODO : mark this condition as manually invalid if this happens
-		c.lg.Errorf("msgCh is not initialized for condition %s", c.conditionName)
-		return
-	}
-	for {
-		select {
-		case <-c.parentCtx.Done():
-			return
-		case <-c.evaluationCtx.Done():
-			return
-		case msg := <-c.msgCh:
-			var status T
-			err := json.Unmarshal(msg.Data, &status)
-			if err != nil {
-				c.lg.Error(err)
-			}
-			healthy, ts := c.alpha(status)
-			incomingState := alertstorage.State{
-				Healthy:   healthy,
-				Firing:    c.IsFiring(),
-				Timestamp: ts,
-			}
-			c.UpdateState(c.evaluationCtx, &incomingState)
-		}
-	}
-}
-
-// infinite & blocking : must be run in a goroutine
-func (c *InternalConditionEvaluator[T]) EvaluateLoop() {
-	defer c.cancelEvaluation() // cancel parent context, if we return (non-recoverable)
-	ticker := time.NewTicker(time.Second * 10)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-c.parentCtx.Done():
-			return
-		case <-c.evaluationCtx.Done():
-			return
-		case <-ticker.C:
-			lastKnownState, err := c.storageNode.GetConditionStatusTracker(c.evaluationCtx, c.conditionId)
-			if err != nil {
-				continue
-			}
-			if !lastKnownState.Healthy {
-				c.lg.Debugf("condition %s is unhealthy", c.conditionName)
-				interval := timestamppb.Now().AsTime().Sub(lastKnownState.Timestamp.AsTime())
-				if interval > c.evaluateDuration {
-					c.lg.Debugf("triggering alert for condition %s", c.conditionName)
-					c.triggerHook(c.evaluationCtx, c.conditionId, metrics.MergeLabels(c.alertmanagerlabels, map[string]string{
-						shared.BackendConditionIdLabel: c.conditionId,
-					}))
-					if err != nil {
-						c.lg.Error(err)
-					}
-					if !c.IsFiring() {
-						c.SetFiring(true)
-						err = c.UpdateState(c.evaluationCtx, &alertstorage.State{
-							Healthy:   lastKnownState.Healthy,
-							Firing:    c.IsFiring(),
-							Timestamp: timestamppb.Now(),
-						})
-						if err != nil {
-							c.lg.Error(err)
-						}
-						err = c.storageNode.OpenInterval(c.evaluationCtx, c.conditionId, timestamppb.Now())
-						if err != nil {
-							c.lg.Error(err)
-						}
-					}
-				} else {
-					c.SetFiring(false)
-				}
-			} else if lastKnownState.Healthy && c.IsFiring() {
-				c.lg.Debugf("condition %s is finally healthy after having fired", c.conditionName)
-				c.SetFiring(false)
-				err = c.storageNode.CloseInterval(c.evaluationCtx, c.conditionId, timestamppb.Now())
-				if err != nil {
-					c.lg.Error(err)
-				}
-			}
-		}
-	}
-}
-
-func (c *InternalConditionEvaluator[T]) SetFiring(firing bool) {
-	c.firingLock.Lock()
-	defer c.firingLock.Unlock()
-	c.inMemoryFiring = firing
-}
-
-func (c *InternalConditionEvaluator[T]) IsFiring() bool {
-	c.firingLock.RLock()
-	defer c.firingLock.RUnlock()
-	return c.inMemoryFiring
-}
-
-func (c *InternalConditionEvaluator[T]) UpdateState(ctx context.Context, s *alertstorage.State) error {
-	c.stateLock.Lock()
-	defer c.stateLock.Unlock()
-	return c.storageNode.UpdateConditionStatusTracker(ctx, c.conditionId, s)
-}
-
-func (c *InternalConditionEvaluator[T]) CalculateInitialState() {
-	incomingState := alertstorage.DefaultState()
-	if st, getErr := c.storageNode.GetConditionStatusTracker(c.evaluationCtx, c.conditionId); getErr != nil {
-		if err := c.storageNode.CreateConditionStatusTracker(c.evaluationCtx, c.conditionId, incomingState); err != nil {
-			// TODO : mark this condition as manually invalid when this happens
-			c.cancelEvaluation()
-			return
-		}
-
-	} else {
-		incomingState = st
-	}
-	if incomingState.Firing { // need to update this in memory value
-		c.SetFiring(true)
-	}
-	c.UpdateState(c.evaluationCtx, incomingState)
-}
-
 func (p *Plugin) onSystemConditionCreate(conditionId, conditionName string, condition *alertingv1.AlertConditionSystem) error {
 	lg := p.Logger.With("onSystemConditionCreate", conditionId)
 	lg.Debugf("received condition update: %v", condition)
@@ -470,7 +322,7 @@ func (p *Plugin) onSystemConditionCreate(conditionId, conditionName string, cond
 		cancelEvaluation:   cancel,
 		evaluateDuration:   condition.GetTimeout().AsDuration(),
 		alertmanagerlabels: map[string]string{},
-		alpha: func(h health.StatusUpdate) (health bool, ts *timestamppb.Timestamp) {
+		healthOnMessage: func(h health.StatusUpdate) (health bool, ts *timestamppb.Timestamp) {
 			return h.Status.Connected, h.Status.Timestamp
 		},
 		triggerHook: func(ctx context.Context, conditionId string, labels map[string]string) {
@@ -521,10 +373,6 @@ func (p *Plugin) onDownstreamCapabilityConditionCreate(conditionId, conditionNam
 		cancel()
 		return err
 	}
-	fireOn := map[string]struct{}{}
-	for _, s := range condition.GetCapabilityState() {
-		fireOn[s] = struct{}{}
-	}
 	lg.Debugf("Creating agent disconnect with timeout %s", condition.GetFor().AsDuration())
 	evaluator := &InternalConditionEvaluator[*corev1.ClusterHealthStatus]{
 		lg:                 lg,
@@ -536,19 +384,22 @@ func (p *Plugin) onDownstreamCapabilityConditionCreate(conditionId, conditionNam
 		cancelEvaluation:   cancel,
 		evaluateDuration:   condition.GetFor().AsDuration(),
 		alertmanagerlabels: map[string]string{},
-		alpha: func(h *corev1.ClusterHealthStatus) (health bool, ts *timestamppb.Timestamp) {
-			healthy := false
-			lg.Debug("found health conditions %v", h.HealthStatus.Health.Conditions)
+		healthOnMessage: func(h *corev1.ClusterHealthStatus) (healthy bool, ts *timestamppb.Timestamp) {
+			healthy = true
+			lg.Debugf("found health conditions %v", h.HealthStatus.Health.Conditions)
 			for _, s := range h.HealthStatus.Health.Conditions {
-				if _, ok := fireOn[s]; !ok {
-					healthy = true
-					break
+				for _, badState := range condition.GetCapabilityState() {
+					if strings.Contains(s, badState) {
+						healthy = false
+						break
+					}
 				}
+
 			}
 			return healthy, h.HealthStatus.Status.Timestamp
 		},
 		triggerHook: func(ctx context.Context, conditionId string, labels map[string]string) {
-			p.TriggerAlerts(ctx, &alertingv1.TriggerAlertsRequest{
+			_, _ = p.TriggerAlerts(ctx, &alertingv1.TriggerAlertsRequest{
 				ConditionId: &corev1.Reference{Id: conditionId},
 				Annotations: labels,
 			})
